+++ build_test/lib/src/written_asset_reader.dart
@@ -27,8 +27,13 @@ class WrittenAssetReader extends AssetReader implements AssetReaderState {
   WrittenAssetReader(this.source, [this.filterSpy]);
 
   @override
-  WrittenAssetReader copyWith({FilesystemCache? cache}) =>
-      WrittenAssetReader(source.copyWith(cache: cache), filterSpy);
+  WrittenAssetReader copyWith({
+    AssetPathProvider? assetPathProvider,
+    FilesystemCache? cache,
+  }) => WrittenAssetReader(
+    source.copyWith(assetPathProvider: assetPathProvider, cache: cache),
+    filterSpy,
+  );
 
   @override
   Filesystem get filesystem => source.filesystem;